+++ flake.nix
@@ -47,4 +47,4 @@
         };
       });
 }
-# nix-direnv cache busting line: sha256-hpFVgsUKswE7g69EieoeKGPR1nVkcRmBhDKbnB2CDBg=
+# nix-direnv cache busting line: sha256-yfOl/gWIijLlqchXFiTRZ7vlgS/kn0xOmv52TFMYs+E=